Successful http request(200 request status code) and unprocessable entity(422 request status code) were tested
